Move the `test-delayed` function from its own file to a common file
Now that the main set of utilities is in a js file and not in an html
file (loaded with HTML imports), we don't need a separate file for it.

This function had to be in a separate js file instead of an html file
with the rest of utilities in order to solve the Firefox issue where
the html-hosted utilities couldn't be used before HTML polyfill finished
initialization.