Allow passing null/undefined entries in the list of checkers
These entries will just be ignored. This is convenient for composing
lists of checkers where some content is rendered conditionally. Before
this, it was only possible to prepare a filtered smaller-size array if
some content is expected to be not rendered conditionally, and now it is
possible to just place null/undefined, which will just be omitted from
the resulting checkers' array.